doc_intelligence/function/requirements.txt
azure-functions
azure-ai-documentintelligence